refactor(backend): move token crud into its own module
Pure move: every function body is byte-identical to its original in
crud/__init__.py, verified by parsing both files with ast and comparing the
unparsed functions.

A transitional re-export at the extraction site keeps callers working so this
commit is reviewable on its own; the next commit migrates them and drops it.
It sits at the extraction site rather than in the import block so the import
header stays byte-identical to main and cannot conflict with the concurrent
comment extraction.

The module-level datetime/timezone import stays in crud/__init__.py — the Task
block still uses it — and is imported fresh in token.py.
